Refactor Sensor to package init and add base support for GPIO Sensors
- Use enum for error codes
- Fix schema retrievers
- misc fixes
