+++ test/mock.py
@@ -2,7 +2,7 @@ import random
 from datetime import datetime
 
 from commons.data import Response
-from sensors.sensor import Sensor
+from sensors import Sensor
 
 
 class MockSensor(Sensor):
